requirements.txt
fastapi
uvicorn
httpx
requests
python-dotenv
requests-toolbelt